move share function handlers into view submit handler
