+++ handlers/ExecuteBlockActionHandler.ts
@@ -183,84 +183,6 @@ export class ExecuteBlockActionHandler {
                 }
                 break;
             }
-            case Modals.SHARE_ACTION: {
-                try{
-                    const persis = this.read.getPersistenceReader();
-                    const association_input = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `${user.id}#share_input`);
-                    const share_record = await persis.readByAssociation(association_input);
-                    if (share_record) {
-                        let input_str = share_record[0]['share_input'] as string
-                        await sendMessage(
-                                this.modify,
-                                room,
-                                user,
-                                input_str,
-                            );
-                    } else {
-                        this.app.getLogger().debug("error: no sharing content");
-                    }
-                } catch(err){
-                    console.log("share error:"+err);
-                }
-                break;
-            }
-            case Modals.SHARE_GITHUB_ACTION: {
-                console.log("SHARE_GITHUB_ACTION");
-                try{
-                    const persis = this.read.getPersistenceReader();
-                    const share_record = await persis.readByAssociation(new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `${user.id}#share_github_input`));
-                    const share_repo_record = await persis.readByAssociation(new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `${user.id}#share_github_repo_input`));
-                    const share_path_record = await persis.readByAssociation(new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `${user.id}#share_github_path_input`));
-                    const share_commit_record = await persis.readByAssociation(new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `${user.id}#share_github_commit_input`));
-                    const share_branch_record = await persis.readByAssociation(new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `${user.id}#share_github_branch_input`));
-                    if (share_record && share_repo_record && share_path_record && share_commit_record) {
-                        let input_str = share_record[0]['share_github_input'] as string
-                        let repo = share_repo_record[0]['share_github_repo_input'] as string
-                        let path = share_path_record[0]['share_github_path_input'] as string
-                        let commit = share_commit_record[0]['share_github_commit_input'] as string
-                        let branch = share_branch_record[0]['share_github_branch_input'] as string
-                        let accessToken = await getAccessTokenForUser(this.read, user, this.app.oauth2Config);
-                        
-                        if (!accessToken) {
-                            await sendNotification(this.read, this.modify, user, room, "Please first login To Github!");
-                        } else {
-                            let getresponse = await getNewCode(this.http, repo, path, input_str, commit, accessToken?.token, branch);
-                            
-                            if (getresponse && !getresponse?.serverError) {
-                                console.log("File Exists!");
-                                await sendNotification(this.read,this.modify,user,room,`File exists, current content will be overwritten!`);
-                                console.log("Inside get sha: "+getresponse.sha)
-                                let response = await uploadNewCode(
-                                    this.http, repo, path, input_str, commit, accessToken?.token, branch, getresponse.sha
-                                )
-                                if(response && !response?.serverError){
-                                    await sendNotification(this.read,this.modify,user,room,`Successfully shared your code to Github!`);
-                                }else{
-                                    await sendNotification(this.read,this.modify,user,room,`Sharing to Github failed!`);
-                                }
-                            }
-                            else {
-                                console.log("File not exist, create new one")
-                                let response = await uploadNewCode(
-                                    this.http, repo, path, input_str, commit, accessToken?.token, branch
-                                )
-                                if(response && !response?.serverError){
-                                    await sendNotification(this.read,this.modify,user,room,`Successfully shared your code to Github!`);
-                                }else{
-                                    await sendNotification(this.read,this.modify,user,room,`Sharing to Github failed!`);
-                                }
-                            }
-                            
-                        }
-                    } else {
-                        await sendNotification(this.read, this.modify, user, room, "You have to enter all information!");
-                    }
-                }
-                catch(err){
-                    console.log("Share Github Error: "+JSON.stringify(err));
-                }
-                break;
-            }
             case Modals.REGEN_BUTTON_ACTION: {
                 if (!room) {
                     this.app.getLogger().error("Room is not specified!");
